requirements.txt
plotly==2.0.10
numpy==1.13.0
airtable==0.3.1
requests==2.18.1
responses==0.5.1
python-dotenv==0.6.4